Prekernel: Move the generate_secure_seed function out of init.cpp
As part of a commit series trying to make the prekernel nicer, we should
move and refactor functions out of the init.cpp file if possible.
