+++ Kernel/Prekernel/Random.h
@@ -1,0 +1,11 @@
+/*
+ * Copyright (c) 2024, Liav A. <liavalb@hotmail.co.il>
+ *
+ * SPDX-License-Identifier: BSD-2-Clause
+ */
+
+#pragma once
+
+#include <AK/Types.h>
+
+u64 generate_secure_seed();
